+++ admin/drivers/pgsql.inc.php
@@ -271,6 +271,10 @@ if (isset($_GET["pgsql"])) {
 				}
 			}
 
+			if (min_version('12', '0', $connection)) {
+				$this->generated = ["STORED"];
+			}
+
 			// No "SQL" to avoid CSRF.
 			$this->operators = [
 				"=", "<", ">", "<=", ">=", "!=",
@@ -535,7 +539,7 @@ ORDER BY a.attnum"
 			if (in_array($row['attidentity'], ['a', 'd'])) {
 				$row['default'] = 'GENERATED ' . ($row['attidentity'] == 'd' ? 'BY DEFAULT' : 'ALWAYS') . ' AS IDENTITY';
 			}
-			$row["generated"] = ($row["attgenerated"] == "s");
+			$row["generated"] = ($row["attgenerated"] == "s" ? "STORED" : "");
 			$row["null"] = !$row["attnotnull"];
 			$row["auto_increment"] = $row['attidentity'] || preg_match('~^nextval\(~i', $row["default"]);
 			$row["privileges"] = ["insert" => 1, "select" => 1, "update" => 1, "where" => 1, "order" => 1];
@@ -667,9 +671,9 @@ ORDER BY conkey, conname") as $row) {
 					}
 					$alter[] = "ALTER $column TYPE$val[1]";
 					$sequence_name = $table . "_" . idf_unescape($val[0]) . "_seq";
-					$alter[] = "ALTER $column " . ($val[3] ? "SET$val[3]"
+					$alter[] = "ALTER $column " . ($val[3] ? "SET" . preg_replace('~GENERATED ALWAYS(.*) STORED~', 'EXPRESSION\1', $val[3])
 						: (isset($val[6]) ? "SET DEFAULT nextval(" . q($sequence_name) . ")"
-						: "DROP DEFAULT"
+						: "DROP DEFAULT" //! change to DROP EXPRESSION with generated columns
 					));
 					if (isset($val[6])) {
 						$sequence = "CREATE SEQUENCE IF NOT EXISTS " . idf_escape($sequence_name) . " OWNED BY " . idf_escape($table) . ".$val[0]";