+++ admin/drivers/sqlite.inc.php
@@ -157,6 +157,10 @@ if (isset($_GET["sqlite"])) {
 				]
 			];
 
+			if (min_version('3.31', '0', $connection)) {
+				$this->generated = ["STORED", "VIRTUAL"];
+			}
+
 			// REGEXP can be a user defined function.
 			$this->operators = [
 				"=", "<", ">", "<=", ">=", "!=",
@@ -316,7 +320,7 @@ if (isset($_GET["sqlite"])) {
 	function fields($table) {
 		$return = [];
 		$primary = "";
-		foreach (get_rows("PRAGMA table_info(" . table($table) . ")") as $row) {
+		foreach (get_rows("PRAGMA table_" . (min_version('3.31') ? "x" : "") . "info(" . table($table) . ")") as $row) {
 			$name = $row["name"];
 			$type = strtolower($row["type"]);
 			$default = $row["dflt_value"];
@@ -339,12 +343,19 @@ if (isset($_GET["sqlite"])) {
 			}
 		}
 		$sql = Connection::get()->getResult("SELECT sql FROM sqlite_master WHERE type = 'table' AND name = " . q($table));
-		preg_match_all('~(("[^"]*+")+|[a-z0-9_]+)\s+text\s+COLLATE\s+(\'[^\']+\'|\S+)~i', $sql, $matches, PREG_SET_ORDER);
+		$idf = '(("[^"]*+")+|[a-z0-9_]+)';
+		preg_match_all('~' . $idf . '\s+text\s+COLLATE\s+(\'[^\']+\'|\S+)~i', $sql, $matches, PREG_SET_ORDER);
 		foreach ($matches as $match) {
 			$name = str_replace('""', '"', preg_replace('~^"|"$~', '', $match[1]));
 			if ($return[$name]) {
 				$return[$name]["collation"] = trim($match[3], "'");
 			}
+		}
+		preg_match_all('~' . $idf . '\s.*GENERATED ALWAYS AS \((.+)\) (STORED|VIRTUAL)~i', $sql, $matches, PREG_SET_ORDER);
+		foreach ($matches as $match) {
+			$name = str_replace('""', '"', preg_replace('~^"|"$~', '', $match[1]));
+			$return[$name]["default"] = $match[3];
+			$return[$name]["generated"] = strtoupper($match[4]);
 		}
 		return $return;
 	}
@@ -614,6 +625,9 @@ if (isset($_GET["sqlite"])) {
 		}
 
 		foreach ($fields as $key => $field) {
+			if (preg_match('~GENERATED~', $field[3])) {
+				unset($originals[array_search($field[0], $originals)]);
+			}
 			$fields[$key] = "  " . implode($field);
 		}